Added reference_python; renamed reference.py to reference_jython.
Fixed up ib/aux/__init__.py to implement all of the expected behavior
of translated code.

Corrected imports and __init__ method in the configuration module for
EReader.py.  Updated ib/ext/EReader.py also included in this checkin.

Changed the method sorter to just an assignment in the configuration
module for EClientSocket.py.